requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http==5.3.4
apache-airflow-providers-postgres==6.4.0
apache-airflow-providers-slack==9.4.0
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-google==19.5.0
astro-sdk-python
google-cloud-storage
pandas==2.3.1
pandas_market_calendars==5.1.3
pendulum==3.1.0
astronomer-cosmos
dbt-core==1.10.15
dbt-postgres==1.9.1
pathlib==1.0.1
requests==2.32.5
